Update CatanBoard.java
Added the generic ports as well as a lot of small things (made players have a reference to the catanboard in their class; fixed the cancel issue when players are using ports; fixed the nesting issue in my usePorts() method). There is still an issue with ownership of indexes that I'm not sure why it's occurring.