Update Port.java
Got rid of the occupied boolean for the Port class; I realized it wasn't necessary because the Port class was a glorified position holder. I could have probably made an interface that Port and Index inherit from, but I made Port after Index, and changing a lot around seems unnecessary at this point.